Moved field validation tests in their own module
